@@ -1,7 +1,6 @@
-// postcss.config.cjs
-module.exports = {
+export default {
   plugins: {
     tailwindcss: {},
     autoprefixer: {},
   },
-};
+}
