@@ -1,44 +1,105 @@
-export const products = [
+export const PRODUCTS = [
   {
     id: 1,
-    name: "Iphone 13 128GB - Blue",
-    price: "$699",
-    image: "/assets/Iphone_blue.png",
-    rate: 3,
+    name: "Titanium Headphones",
+    tagline: "Som puro. Silêncio absoluto.",
+    price: 3499,
+    image: "https://images.unsplash.com/photo-1505740420928-5e560c06d30e?auto=format&fit=crop&q=80&w=1000",
+    description: "A reinvenção da acústica. Com drivers de titânio de alta fidelidade e cancelamento de ruído adaptativo que silencia o mundo para que você ouça apenas a música. Design unibody em alumínio aeroespacial.",
+    features: ["Áudio Espacial", "20h de Bateria", "Cancelamento Ativo"]
   },
   {
     id: 2,
-    name: "Iphone 13 128GB - Midnight",
-    price: "$749",
-    image: "/assets/Iphone_meianoite.png",
-    rate: 4,
+    name: "Watch Ultra",
+    tagline: "Aventura em cada pulso.",
+    price: 5299,
+    image: "https://images.unsplash.com/photo-1523275335684-37898b6baf30?auto=format&fit=crop&q=80&w=1000",
+    description: "Construído para resistir aos elementos. Caixa de titânio robusta, GPS de dupla frequência e até 36 horas de bateria. A ferramenta definitiva para atletas e exploradores.",
+    features: ["Caixa de Titânio", "GPS Duplo", "Mergulho até 100m"]
   },
   {
     id: 3,
-    name: "Iphone 13 512GB - Red",
-    price: "$999",
-    image: "/assets/Iphone_red.png",
-    rate: 5,
+    name: "Studio Display",
+    tagline: "Veja o que você pode fazer.",
+    price: 8799,
+    image: "https://images.unsplash.com/photo-1527443224154-c4a3942d3acf?auto=format&fit=crop&q=80&w=1000",
+    description: "Uma janela imersiva para o seu trabalho. Cores P3, 5K de resolução e câmera ultra-wide integrada. O parceiro perfeito para o seu fluxo criativo.",
+    features: ["Resolução 5K", "Câmera 12MP", "Áudio Espacial"]
   },
   {
     id: 4,
-    name: "Iphone 13 256GB - Rose",
-    price: "$849",
-    image: "/assets/Iphone_rose.png",
-    rate: 0,
+    name: "Pro Keyboard",
+    tagline: "Digite o futuro.",
+    price: 1299,
+    image: "https://images.unsplash.com/photo-1587829741301-dc798b91ddce?auto=format&fit=crop&q=80&w=1000",
+    description: "Mecanismo tesoura refinado para uma experiência de digitação silenciosa e precisa. Retroiluminação inteligente que se adapta ao seu ambiente.",
+    features: ["Touch ID", "Wireless", "Bateria de 1 mês"]
+  }
+];
+
+export const ACCESSORIES = [
+  {
+    id: 101,
+    name: "AirPods Max",
+    category: "audio",
+    price: "R$ 6.590",
+    image: "https://images.unsplash.com/photo-1613040809024-b4ef7ba99bc3?auto=format&fit=crop&q=80&w=800",
+    color: "Silver"
   },
   {
-    id: 5,
-    name: "Iphone 13 128GB - Stelar",
-    price: "$719",
-    image: "/assets/Iphone_stelar.png",
-    rate: 4,
+    id: 102,
+    name: "MagSafe Charger",
+    category: "power",
+    price: "R$ 499",
+    image: "https://images.unsplash.com/photo-1615526675159-e248c3021d3f?auto=format&fit=crop&q=80&w=800",
+    color: "White"
   },
   {
-    id: 6,
-    name: "Iphone 13 256GB - Blue",
-    price: "$769",
-    image: "/assets/Iphone_blue.png",
-    rate: 0,
+    id: 103,
+    name: "Capa Silicone iPhone 15",
+    category: "protection",
+    price: "R$ 599",
+    image: "https://images.tcdn.com.br/img/img_prod/1249456/capa_iphone_15_plus_de_silicone_aveludada_colorida_8003_1_26f1d2e5d4c7c3ebeb6819fa61c57aa4.png",
+    color: "Midnight"
   },
+  {
+    id: 104,
+    name: "AirPods Pro (2ª Ger)",
+    category: "audio",
+    price: "R$ 2.599",
+    image: "https://images.unsplash.com/photo-1600294037681-c80b4cb5b434?auto=format&fit=crop&q=80&w=800",
+    color: "White"
+  },
+  {
+    id: 105,
+    name: "Pulseira Alpine Loop",
+    category: "wearables",
+    price: "R$ 999",
+    image: "https://images.unsplash.com/photo-1523275335684-37898b6baf30?auto=format&fit=crop&q=80&w=800",
+    color: "Orange"
+  },
+  {
+    id: 106,
+    name: "Adaptador USB-C 20W",
+    category: "power",
+    price: "R$ 219",
+    image: "https://m.media-amazon.com/images/I/31p-x70so+L._AC_UF1000,1000_QL80_.jpg",
+    color: "White"
+  },
+  {
+    id: 107,
+    name: "HomePod mini",
+    category: "audio",
+    price: "R$ 999",
+    image: "https://images.unsplash.com/photo-1544244015-0df4b3ffc6b0?auto=format&fit=crop&q=80&w=800",
+    color: "Space Gray"
+  },
+  {
+    id: 108,
+    name: "Carteira MagSafe",
+    category: "protection",
+    price: "R$ 720",
+    image: "https://m.media-amazon.com/images/I/714BOrCsK+L.jpg",
+    color: "Leather"
+  }
 ];
